examples/official/9.x/flet_chat_gemini/requirements.txt
google-generativeai
dbr
flet 